+++ src/main/groovy/com/ing/scripts/ScriptCreateDB.groovy
@@ -8,7 +8,7 @@ def sql = Sql.newInstance(url, user, password, driver)
 
 // ... create 'sql' instance
 sql.execute '''
-  CREATE TABLE Users (
+  CREATE TABLE TB_Users (
     id          INTEGER GENERATED BY DEFAULT AS IDENTITY,
     user_name   VARCHAR(64),
     user_password    VARCHAR(64),
@@ -16,17 +16,5 @@ sql.execute '''
   );
 '''
 // close 'sql' instance ...
-
-sql.execute "INSERT INTO Users(user_name, user_password, version) VALUES ('admin', 'admin', 0)"
-sql.execute "INSERT INTO Users(user_name, user_password, version) VALUES ('user', 'user', 0)"
-
-def rowNum = 0
-sql.query('SELECT user_name, user_password FROM Users') { resultSet ->
-    while (resultSet.next()) {
-        def user_name  = resultSet.getString('user_name')
-        def user_password = resultSet.getString('user_password')
-        println "$user_name  $user_password"
-    }
-}
 
 sql.close()
